requirements.txt
streamlit
torch
torchvision
timm
open-clip-torch
opencv-python
matplotlib
umap-learn
Pillow